Remove AnalysisMetaData deps where schemas is not required
Many components used AnalysisMetaData but didn't need all attributes.
This pulled in a lot of dependencies. Especially the Schemas interface
is bad. The only implementation pulls in stuff like the ThreadPool and
ClusterService.

By being concrete about the actual dependencies it becomes a bit more
troublesome to pass around all the different objects but the
dependencies become more explicit and testing should get a lot easier.